requirements.txt
selenium==3.141.0
requests==2.22.0
pytest==5.0.1
PyYAML==5.1.2
python-dateutil==2.8.1